Added enemy damage text, both player and enemy go down relative to their damage, enemy damage scales really slowly using log(level, 2f)
heal button in top left to heal player by 1k hp

need to set up a coroutine for the attack method, as currently it just reduces the damage based on every frame, need it to reduce based on a set scale. e.g. users attack speed and a default enemy attack speed which we can add in, so two coroutines actually

one for the enemy's attack and one for the players attack. both will need stopping when one of the other dies.