Updated animations to be a bit smoother, also set animation for death, which the destruction of the enemy unit waits for the length of time for the death animation, then destroys it.
need to set a heal animation, but have it as a layered animation, so that it does not directly affect the animation that is currently happening, e.g. the green heals stuff goes over the top of the current animation.

could do with some neatening up, but it works okay for now